refactor sql to new file
